+++ terrain/instancing/load_instance_block_task.cpp
@@ -12,28 +12,29 @@
 
 namespace zylann::voxel {
 
-LoadInstanceChunkTask::LoadInstanceChunkTask( //
-		std::shared_ptr<InstancerTaskOutputQueue> output_queue, //
-		Ref<VoxelStream> stream, //
+LoadInstanceChunkTask::LoadInstanceChunkTask(
+		std::shared_ptr<InstancerTaskOutputQueue> output_queue,
+		Ref<VoxelStream> stream,
+		Ref<VoxelGenerator> voxel_generator,
 		std::shared_ptr<InstancerQuickReloadingCache> quick_reload_cache,
-		Ref<VoxelInstanceLibrary> library, //
-		Array mesh_arrays, //
-		Vector3i grid_position, //
-		uint8_t lod_index, //
-		uint8_t instance_block_size, //
-		uint8_t data_block_size, //
-		UpMode up_mode //
+		Ref<VoxelInstanceLibrary> library,
+		Array mesh_arrays,
+		Vector3i grid_position,
+		uint8_t lod_index,
+		uint8_t instance_block_size,
+		uint8_t data_block_size,
+		UpMode up_mode
 ) :
-		//
-		_output_queue(output_queue), //
-		_stream(stream), //
-		_quick_reload_cache(quick_reload_cache), //
-		_library(library), //
-		_mesh_arrays(mesh_arrays), //
-		_render_grid_position(grid_position), //
-		_lod_index(lod_index), //
-		_instance_block_size(instance_block_size), //
-		_data_block_size(data_block_size), //
+		_output_queue(output_queue),
+		_stream(stream),
+		_voxel_generator(voxel_generator),
+		_quick_reload_cache(quick_reload_cache),
+		_library(library),
+		_mesh_arrays(mesh_arrays),
+		_render_grid_position(grid_position),
+		_lod_index(lod_index),
+		_instance_block_size(instance_block_size),
+		_data_block_size(data_block_size),
 		_up_mode(up_mode) //
 {
 #ifdef DEBUG_ENABLED
@@ -237,6 +238,7 @@ void LoadInstanceChunkTask::run(ThreadedTaskContext &ctx) {
 						task->up_mode = _up_mode;
 						task->surface_arrays = _mesh_arrays;
 						task->generator = item.generator;
+						task->voxel_generator = _voxel_generator;
 						task->transforms = std::move(layer.transforms);
 						task->output_queue = _output_queue;